refactor: API 구조 변경 및 인증 API 제거
