+++ src/v1/apis/users.schema.ts
@@ -1,10 +1,10 @@
-import { createResponseSchema } from '../../common/schema/core.schema.js';
+import { createResponseSchema } from '../common/schema/core.schema.js';
 import { z } from 'zod';
 
 export const FindUserResponseSchema = createResponseSchema(
   z.object({
     id: z.number(),
-    name: z.string(),
+    nickname: z.string(),
     email: z.string(),
     createdAt: z.date(),
     updatedAt: z.date(),
